feat: check for duplicates in checkFirmwareVersions
